Support unparse with symbol annotations.
Original merge commit: fcca042a10b3908ffb72d8cf7ae86c71cffdfc6d

######

Add -funparse-with-symbols to enable new behavior


Original commit: 7ad44e7e33de7020a94378b8f98395e7d5edef51
-------

Support unparse with symbol annotations.
When `-fdebug-dump-symbols` is supplied with `-funparse`, include symbol
information in comments in the Fortran output. This will be used for
testing to verify that correct symbols are defined and references in
the right place.

In `UnparseWithSymbols()`, walk the parse tree and collect symbol
definitions and references, organized by statement. When a symbol is
defined across several statement it is associated with the first.
The definition of implicitly defined symbols is associated with the
first reference.

To write out the symbol information, a new optional argument is added to
`Unparse()`: it is a function that is called immediately before each
statement is unparsed. We pass in a function that prints out the symbol
information collected for that statement.

Add `Symbol::GetType()` to make it easier to write the symbol types
and add `Symbol::SetType()` for uniformity.


Original commit: 2e827de4ad23d036fa3436382c8a878a991f11f5
-------