Fix bug in rewriting function reference
Original merge commit: 22f25eadc7e0849bba3601a53ac331903ef5c203

######

Fix bug in rewriting function reference
`a(i)` is parsed as a function reference and needs to be converted to an
array element reference when `a` is an object entity. That determination
was wrong if the symbol for `a` was a symbol representing host-association
or use-association. In that case we need to get to the original symbol
by calling `GetUltimate()` on the symbol.

This was causing symbol09.f90 to get a compilation error because an
array element reference looked like a call to a non-pure function, which
is prohibited inside a DO CONCURRENT.


Original commit: 221e6c52c5460edfbec28904ebe1774ff787fc45
-------